refactor(App, CartManager, MenuItem, Kiosk, KioskUI): 가격 타입을 BigDecimal로 통합

+++ app/src/main/java/kiosk/models/MenuItem.java
@@ -1,10 +1,12 @@
 package kiosk.models;
 
+import java.math.BigDecimal;
+
 import kiosk.enums.MenuCategory;
 
 public record MenuItem(
         String name, // 이름
-        double price, // 가격
+        BigDecimal price, // 가격
         String description, // 설명
         MenuCategory category // 카테고리
 ) {
